+++ tests/testthat/test-system_requirements.R
@@ -149,6 +149,22 @@ test_that("system_requirements return the system requirements if 2nd order depen
   )
 })
 
+test_that("system_requirements errors if DESCRIPTION cannot be parsed", {
+  skip_on_cran()
+  skip_if_offline()
+
+  pkg <- tempfile()
+  dir.create(pkg)
+  on.exit(unlink(pkg, recursive = TRUE))
+
+  writeLines("Not parsable", file.path(pkg, "DESCRIPTION"))
+
+  expect_error(
+    system_requirements("ubuntu", "16.04", pkg),
+    "Could not parse DESCRIPTION", fixed = TRUE
+  )
+})
+
 test_that("system_requirements work with package arguments", {
   skip_on_cran()
   skip_if_offline()
@@ -158,6 +174,42 @@ test_that("system_requirements work with package arguments", {
     c("apt-get install -y libcurl4-openssl-dev",
       "apt-get install -y libssl-dev"
     )
+  )
+})
+
+test_that("system_requirements work with vector package arguments", {
+  skip_on_cran()
+  skip_if_offline()
+
+  expect_equal(
+    sort(system_requirements("ubuntu", "16.04", package = c("curl", "git2r"))),
+
+    sort(c("apt-get install -y software-properties-common",
+      "add-apt-repository -y ppa:cran/libgit2",  "apt-get update",
+      "apt-get install -y libgit2-dev", "apt-get install -y libssh2-1-dev",
+      "apt-get install -y libssl-dev", "apt-get install -y zlib1g-dev",
+      "apt-get install -y libcurl4-openssl-dev"
+    ))
+  )
+})
+
+test_that("system_requirements errors for a single non-existing CRAN package", {
+  skip_on_cran()
+  skip_if_offline()
+
+  expect_error(
+    system_requirements("ubuntu", "16.04", package = c("iDontExist")),
+    "Could not locate package 'iDontExist'", fixed = TRUE
+  )
+})
+
+test_that("system_requirements errors for multiple packages when one does not exist", {
+  skip_on_cran()
+  skip_if_offline()
+
+  expect_error(
+    system_requirements("ubuntu", "16.04", package = c("curl", "iDontExist")),
+    "Could not locate package 'iDontExist'", fixed = TRUE
   )
 })